HBP-95: Small improvements and corrected dependencies
