+++ framework_tvb/tvb/tests/framework/adapters/uploaders/encrypt_decrypt_test.py
@@ -41,8 +41,10 @@ from cryptography.hazmat.primitives.asymmetric import rsa
 from cryptography.hazmat.backends import default_backend
 from tvb.adapters.uploaders.zip_connectivity_importer import ZIPConnectivityImporterModel
 from tvb.core.adapters.abcuploader import ABCUploader
+from tvb.core.adapters.constants import ENCRYPTED_PASSWORD_NAME, ENCRYPTED_DATA_SUFFIX, DECRYPTED_DATA_SUFFIX
+from tvb.core.services.encryption_handler import EncryptionHandler
 from tvb.interfaces.command.demos.importers.encrypt_data import encrypt_password, save_encrypted_password, \
-    ENCRYPTED_PASSWORD_NAME, get_path_to_encrypt, ENCRYPTED_DATA_SUFFIX, DECRYPTED_DATA_SUFFIX, generate_random_password
+    get_path_to_encrypt
 from tvb.tests.framework.core.base_testcase import TransactionalTestCase
 from tvb.basic.profile import TvbProfile
 
@@ -83,7 +85,7 @@ class TestEncryptionDecryption(TransactionalTestCase):
 
         # Generate password
         pass_size = TvbProfile.current.hpc.CRYPT_PASS_SIZE
-        password = generate_random_password(pass_size)
+        password = EncryptionHandler.generate_random_password(pass_size)
 
         # Encrypt files using an AES symmetric key
         encrypted_file_path = get_path_to_encrypt(path_to_file)
